test/requirements.txt
allure-pytest==2.15.3
allure-python-commons==2.15.3
pytest==9.0.2
requests==2.32.5
selenium==4.40.0